requirements.txt
-f /usr/share/pip-wheels
asn1crypto==0.24.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
httplib2==0.11.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
oauth2client==4.1.2
openapi-codec==1.3.2
polyline==1.3.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
requests==2.19.1
rsa==3.4.2
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
